+++ src/GV/ch/grabba.c
@@ -211,7 +211,7 @@ void chGrabba_update(Actor *this){
                     D_80391A80 = this->state;
                     func_802BB3DC(0, 14.0f, 0.92f);
                     __spawnQueue_add_4((GenFunction_4)spawnQueue_actor_f32, 0x11f,
-                        reinterpret_cast(s32, this->position_x), reinterpret_cast(s32, this->position_y), reinterpret_cast(s32, this->position_z)
+                        reinterpret_cast(intptr_t, this->position_x), reinterpret_cast(intptr_t, this->position_y), reinterpret_cast(intptr_t, this->position_z)
                     );
                 }
                 break;
