+++ src/SM/ch/smbottles.c
@@ -490,7 +490,7 @@ void chSmBottles_update(Actor *this) {
     }//L80389B20
 
     if (!this->volatile_initialized) {
-        __spawnQueue_add_1((GenFunction_1) __chSmBottles_spawnMolehill, reinterpret_cast(s32, this->marker));
+        __spawnQueue_add_1((GenFunction_1) __chSmBottles_spawnMolehill, reinterpret_cast(intptr_t, this->marker));
         this->volatile_initialized = TRUE;
     }//L80389B4C
